Price is now a float, qty is back to integer

+++ db/migrate/20160824143332_remove_qty_from_inventories.rb
@@ -1,5 +1,5 @@
 class RemoveQtyFromInventories < ActiveRecord::Migration
   def change
-    remove_column :inventories, :qty, :integer
+    remove_column :inventories, :qty, :float
   end
 end
